requirements.txt
motor
dill
python-dotenv
fastapi
pydantic==1.10.18
pymongo
jsonpickle
gunicorn
uvicorn
rocketpy
uptrace
opentelemetry.instrumentation.fastapi
opentelemetry.instrumentation.requests
opentelemetry-api
opentelemetry-sdk
tenacity
